refactor: markdowntToHtml parser
